requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
websocket-client==1.7.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
